Improving code speed.
git-svn-id: https://svn.code.sf.net/p/redqueenchess/code/redqueen@128 865e88f3-cc1b-4e5f-86d1-2a7dcd23a5c5

+++ src/SearchAgent.h
@@ -34,12 +34,13 @@
 #include "Board.h"
 #include "SimplePVSearch.h"
 #include "TranspositionTable.h"
+#include "StringUtil.h"
 
 static const std::string mainHashName 		= "DefaultHashTable";
 static const size_t defaultDepth			= 5;
 static const size_t defaultHashSize			= 64;
-static const size_t defaultGameSize			= 40;
-static const size_t defaultGameSizeInc		= 5;
+static const int defaultGameSize			= 40;
+static const int defaultGameSizeInc		= 5;
 
 class SearchAgent {
 public:
